Reorder difference table column types
