+++ taxcalc/tbi/__init__.py
@@ -1,4 +1,3 @@
 from taxcalc.tbi.tbi import (run_nth_year_tax_calc_model,
                              run_nth_year_gdp_elast_model,
-                             create_dict_table,
                              reform_warnings_errors)
